+++ src/main/java/de/gematik/test/tiger/validator/model/TestReport.java
@@ -19,11 +19,12 @@ package de.gematik.test.tiger.validator.model;
 import de.gematik.test.tiger.validator.ReportValidationException;
 import io.cucumber.cienvironment.internal.com.eclipsesource.json.Json;
 import io.cucumber.cienvironment.internal.com.eclipsesource.json.JsonObject;
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.*;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 import lombok.Getter;
@@ -34,45 +35,56 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class TestReport {
   /**
-   * Map containing all result data from parsed json files. Key of the map is concatenation of
-   * feature file name (without path and extension) and scenario name
+   * Map containing all result data from parsed JSON files.
+   * Key of the map is a concatenation of feature file name (without path and extension) and scenario name
    */
   Map<String, ScenarioResult> scenarioResults = new HashMap<>();
 
   public void parseReportFromTitus(ZipInputStream reportStream) {
     try {
       ZipEntry entry;
-      while ((entry = reportStream.getNextEntry()) != null) {
-        if (entry.getName().endsWith(".json")) {
+      while ((entry = reportStream.getNextEntry()) != null) { //NOSONAR - security checks are done on titus side
+        if (entry.getName().endsWith(".json")
+            && !(entry.getName().startsWith("bootstrap") || entry.getName().startsWith("nivo"))) {
           log.info("Parsing zip report file '{}'", entry.getName());
-          parseScenarioJsonResult(new String(reportStream.readAllBytes()));
+          parseScenarioJsonResult(new String(reportStream.readAllBytes()), entry.getName());
         }
       }
     } catch (IOException ioe) {
-      throw new ReportValidationException("Error reading zip file", ioe);
+      throw new ReportValidationException(ReportValidationException.MessageId.ERR_READ_ZIP, ioe);
     }
   }
 
+  /**
+   * Parses all JSON files in the given folder (NO RECURSION) and adds the result data to the {@link
+   * #scenarioResults} map.
+   *
+   * @param folder path to the folder containing the JSON files
+   * @throws IOException if the folder does not exist or is not readable
+   */
   public void parseReport(String folder) throws IOException {
-    Files.walkFileTree(
-        Paths.get(folder),
-        new SimpleFileVisitor<>() {
-          @Override
-          public FileVisitResult visitFile(Path path, BasicFileAttributes basicFileAttributes)
-              throws IOException {
-            if (path.toFile().getName().endsWith(".json")) {
-              log.info("Parsing report file '{}'", path.toFile().getAbsolutePath());
-              parseScenarioJsonResult(Files.readString(path));
-            }
-            return FileVisitResult.CONTINUE;
-          }
-        });
+    File f = Paths.get(folder).toFile();
+    if (!f.exists() || !f.isDirectory()) {
+      throw new IOException("Folder '" + folder + "' does not exist or is not a directory");
+    }
+    for (File file : Objects.requireNonNull(f.listFiles())) {
+      if (file.getName().endsWith(".json")) {
+        log.info("Parsing report file '{}'", file.getAbsolutePath());
+        parseScenarioJsonResult(Paths.get(file.toURI()));
+      }
+    }
     log.info("Parsed {} scenario test results", scenarioResults.size());
   }
 
-  public void parseScenarioJsonResult(String jsonContent) {
-    JsonObject json = Json.parse(jsonContent).asObject();
-      ScenarioResult scenario = ScenarioResult.createResult(json);
-      scenarioResults.put(scenario.getFilename() + scenario.getScenarioName(), scenario);
+  public void parseScenarioJsonResult(Path jsonPath) throws IOException {
+    JsonObject json = Json.parse(Files.readString(jsonPath)).asObject();
+    ScenarioResult scenario = ScenarioResult.createResult(json, jsonPath);
+    scenarioResults.put(scenario.getFilename() + scenario.getScenarioName(), scenario);
+  }
+
+  public void parseScenarioJsonResult(String json, String entryName) {
+    ScenarioResult scenario =
+        ScenarioResult.createResult(Json.parse(json).asObject(), Path.of("ZIP Archiv", entryName));
+    scenarioResults.put(scenario.getFilename() + scenario.getScenarioName(), scenario);
   }
 }