+++ ReleaseNotes.md
@@ -1,3 +1,9 @@
+# Release 1.0.6
+
+- exceptions are now end-user-readable (and in german)
+- test coverage improved
+- added copyright headers
+
 # Release 1.0.1
 
 - release on github